requirements.txt
asgiref==3.9.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
